Moved offers classes into offers package
