+++ app/src/test/java/uk/co/jamiecruwys/cart/DiscountedBasketUnitTest.java
@@ -5,8 +5,8 @@ import org.junit.Test;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 
+import uk.co.jamiecruwys.cart.offers.ApplesBuyOneGetOneFreeOffer;
 import uk.co.jamiecruwys.cart.products.Apple;
-import uk.co.jamiecruwys.cart.products.Orange;
 
 import static org.junit.Assert.assertEquals;
